Ensure that the parent property is an int as per the JSON schema
Previously, the parent was a string which violates the schema, this fix gets the ID of the parent instead of the slug. This also assumes that the parent has been processed before the child. This seems to be an assumption that has also been made in the existing wordpress-importer so for now it shouldn't lead to a different outcome.